VISITOR: remove the tedious `ExprFixpoint`
